Added hook for items
